[TASK] Provide TokenServiceInterface as replacement for tokens provided in DI (#618)
Previously, Louketo proxy provided an X-Auth header that was consumed by
`TokenService` in the package `t3g/symfony-datahub-bundle` and passed
via DI to `DataHubClient`.

With the ongoing removal of the Louketo proxy, the header is not present
anymore, leading to issues as an empty string would've been passed as
token, breaking Datahub requests executed by regular users. To overcome
this, the DataHubClient accepts both, an instance of
`TokenServiceInterface` or a string, for now. Using a string is
deprecated.