[TASK] Upgrade to Symfony 6.4 (#683)
With this commit, Datahub is finally upgraded to Symfony 6.4. All
Symfony packages have been upgraded, along with the following key
changes:

UUID
----

The previously used UUID implementation provided by `ramsey/uuid` and 
its related packages is replaced by `symfony/uid` in order to get a 
tighter coupling to Symfony. Naturally, Symfony wants to store UUIDs as
a binary value, which is not suitable for our use-cases as we need to 
search for specific UUIDs in the database via MySQL CLI. Therefore, a 
custom UUID Doctrine type is implemented that forces to write UUIDs as 
plain string values.

Doctrine
--------

All `doctrine/*` packages are updated to their respective latest
version. The upgrade of `doctrine/dbal` to v4 brought a major change how
`\DateTimeInterface` values are treated. It is now **mandatory** to set
the correct instance type, either `\DateTime` or `\DateTimeImmutable`
and use it consistently.

OpenAPI
-------

All OpenAPI annotations are migrated to PHP-native attributes. Since
there is no functional Rector configuration available, this tedious
migration was done manually.

Message Handling
----------------

Since Symfony 6.2 it's possible to configure rate limiting to a message
queue. This comes in handy for our Pretix integration as they have a
rate limit we're exceeding on a regular basis. A new message queue
`async_pretix` is established that now receives all messages
implementing `VoucherCodeMessageInterface`, having a rate limiter
configured according to Pretix's documentation.